revert: optionally refer to commit in the "reference" format
A typical "git revert" commit uses the full title of the original
commit in its title, and starts its body of the message like:

    This reverts commit 8fa7f667cf61386257c00d6e954855cc3215ae91.

This does not encourage the best practice of describing not just
"what" (i.e. "Revert X" on the title says what we did) but "why".
We can instead phrase this first line to be more like

    This reverts commit 8fa7f667 (do this and that, 2022-04-25)

so that the title does not have to be

    Revert "do this and that"

but can be used to describe "why" we are reverting the original
commit.

Introduce the revert.reference configuration variable, which
defaults to false, to allow us to do so.

Signed-off-by: Junio C Hamano <gitster@pobox.com>